Added Granular CMAKE Options Added separate CMAKE options to include each of the individual examples as well as one global option to include all examples.
I set build all examples to OFF as a default since downloading all of SDL 2 and 3 seems excessive.  Instead, the demo from the video and the website are turned on by default.

A note about these options should probably be added to the ReadMe as well, but I wasn't sure where the logical place to put it would be.